[TypeInfo] ArrayShape can resolve key type as callable instead of string
In PHP, array key type is either string or int. Although callable is not entirely false as long as it's a string callable, using callable as a key type for an array seems wrong, and it's unlikely to be what we expect.
This is peculiarly true in environment when key name might collide with global function name.
Proposed change is to check for int/string instead of resolving type.
Because the input is an array, the key is already narrowed down to string|int in the input

This bugfix is also a change of behavior, as callable was likely to be fed in keyTypes, and it won't longer occur with that change.