[DependencyInjection] Ensure deprecation detection does not trigger a PHP error
The `class_exists` check and the condition `$definition instanceof ChildDefinition`
have been swapped with #61215 in order to produce an error if a service is not
available.

This error was later relaxed to a deprecation in #61270, but the executing order
of the condition kept to be `class_exists` first, then `instanceof`.

That means, if `class_exists` fails with a PHP error
(like PHP Fatal Error: Interface "My\Vendor\MyInterface" not found`)
affected uses will not receive the deprecation, but a fatal error instead.